requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
audioop-lts==0.2.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromadb==1.0.15
click==8.1.8
cmake==3.31.4
colorama==0.4.6
coloredlogs==15.0.1
cssselect2==0.8.0
distro==1.9.0
docopt==0.6.2
durationpy==0.10
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.5.0
fsspec==2025.5.1
future==1.0.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
Js2Py==0.74
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.47.1
numpy==2.2.3
oauthlib==3.3.1
onnxruntime==1.22.0
openai==0.28.0
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pefile==2023.2.7
pillow==11.1.0
pipwin==0.5.2
playsound==1.2.2
plotly==6.2.0
posthog==5.4.0
propcache==0.3.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyjsparser==2.7.1
PyJWT==2.10.1
pyphen==0.17.2
PyPika==0.48.9
PyPrind==2.11.3
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pySmartDL==1.3.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzk==0.9
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
setuptools==75.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tk==0.1.0
tokenizers==0.21.2
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.35.0
watchfiles==1.1.0
weasyprint==65.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.23.0
zopfli==0.2.3.post1
